extracted choose member questions and answer context into separate components for wizard, set up render to progress, so correct step shows in the wizard,
